Changed the name of the object in a single object response into results
The main item in the json response was always "results", even if a single object is returned, which is confusing. It might be better to name the item according to the object type, e.g. "location", "timeseries", et ceterat, but that might be more work to implement than using the same itemname for all 'single-object responses'. Discussed with @geriwolters, and we decided that the most practical is to name it "result".